move customer selector to utilities
